% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/forceGraph.R
\name{forceNetwork.igraph}
\alias{forceNetwork.igraph}
\title{A forceNetwork of an igraph object}
\usage{
forceNetwork.igraph(g, weight, group)
}
\arguments{
\item{weight}{The graph edge attribute used for weighting the graph.}

\item{group}{The vertex attribute used for coloring the graph.}

\item{graph}{An instance of class \code{igraph}}
}
\description{
This function returns a javascript/html object of d3 force network
}

